refactor: Remove support for changing Config
For the previous few commits, since the introduction of a custom `Mapping`
implementation for `Config`, it has had to be writable to support
disabling of extensions that failed their self-check late in the startup
sequence.

After the recent startup sequence refactoring, the `Config`` type is
only created after extension disabling is done, and the `Config`` can thus
be read-only again, matching the previous behavior, implemented using the
`ReadOnlyDict` class.